requirements.txt
attrs==17.3.0
bleach==2.1.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.4.2
decorator==4.1.2
Django==2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
entrypoints==0.2.3
Flask==0.12.2
freesms==0.1.2
future==0.16.0
html5lib==1.0.1
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
oauthlib==2.0.6
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
py==1.5.2
Pygments==2.2.0
PyJWT==1.5.3
PySocks==1.6.7
pytest==3.3.1
pytest-cov==2.5.1
python-dateutil==2.6.1
python-twitter==3.3
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
twilio==6.10.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.8
